fix null vehicle-id in sql

+++ backend/src/main/sql/fintraffic/efti/backend/db/consignment.sql
@@ -10,7 +10,7 @@ where consignment.gate_url = :gate-url and
 -- :name select-consignments
 select consignment.* from consignment_json consignment
 where
-  (:vehicle-id is null or exists (
+  (:vehicle-id::int is null or exists (
     select from transport_vehicle vehicle
     where vehicle.consignment_id = consignment.id and vehicle.vehicle_id = :vehicle-id))
 limit :limit offset :offset;
